requirements.txt
python-dotenv==0.20.0
wheel-0.43.0
setuptools-69.5.1
sentence-transformers==2.7.0
numpy
pandas==1.5.3
regex
scrapy-playwright==0.0.26
pytest-playwright==0.3.2
scrapy==2.8.0
nltk==3.8.1
tqdm
bs4==0.0.1
urllib3==1.26.6
spacy==3.7.4
chromadb==0.4.24
langchain==0.1.16
chardet==5.2.0